+++ dvgutils/modules/image_capture.py
@@ -2,25 +2,32 @@ import os
 import cv2
 
 from ..fs import list_files
+from .transform import Transform
 
 
 class ImageCapture:
+    def __init__(self, conf):
+        self.conf = conf
+        self.path = conf["path"]
+        self.valid_ext = conf["valid_ext"]
+        self.contains = conf["contains"] if "contains" in conf else None
+        self.level = conf["level"] if "level" in conf else None
 
-    def __init__(self, path, valid_exts=(".jpg", ".png"), contains=None, level=None):
-        self.path = path
-        self.valid_exts = valid_exts
-        self.contains = contains
-        self.level = level
+        # Setup optional image transformation function (resizing, flipping, etc.)
+        self.transform = Transform(conf["transform"]) if "transform" in conf else None
 
         if os.path.isfile(self.path):
             self.source = iter([self.path])
         else:
-            self.source = list_files(self.path, self.valid_exts, self.contains, self.level)
+            self.source = list_files(self.path, self.valid_ext, self.contains, self.level)
 
     def read(self):
         try:
             filename = next(self.source)
             image = cv2.imread(filename)
+            if self.transform:
+                image = self.transform(image)
             return filename, image
         except StopIteration:
             return None, None
+