+++ examples/capture_image.py
@@ -1,22 +1,20 @@
 import os
 import logging
 
-from dvgutils import setup_logger
+from dvgutils import setup_logger, load_config
 from dvgutils.modules import ImageCapture, ShowImage, SaveImage, Metrics, Progress
 from dvgutils.pipeline import CaptureImagePipe, Pipeline, ShowImagePipe, SaveImagePipe, MetricsPipe, ProgressPipe
 
 from utils.vis import visualize_image_info
 
+
 def parse_args():
     import argparse
 
     parser = argparse.ArgumentParser()
-    parser.add_argument("-i", "--input", required=True, type=str,
-                        help="image file or images input path")
-    parser.add_argument("-ie", "--image-ext", default="jpg", choices=["jpg", "png"],
-                        help="image extension (default: jpg)")
-    parser.add_argument("-c", "--contains", type=str,
-                        help="image name should contain given string")
+    parser.add_argument("-cf", "--conf", default="config/capture_image.yml",
+                        help="Path to the input configuration file (default: config/capture_image.yml)")
+    parser.add_argument("-cfo", "--conf-overwrites", nargs="+", type=str)
     parser.add_argument("-o", "--output", type=str,
                         help="path to output directory")
     parser.add_argument("--no-display", dest='display', action="store_false",
@@ -30,11 +28,14 @@ def parse_args():
 
 
 def capture_image(args):
+    # Setup logger
     logger = logging.getLogger(__name__)
+    # Read configuration
+    conf = load_config(args["conf"], args["conf_overwrites"])
 
     # Setup processing modules
-    image_capture = ImageCapture(args["input"], "." + args["image_ext"], args["contains"])
-    save_image = SaveImage(args["output"], args["image_ext"]) if args["output"] else None
+    image_capture = ImageCapture(conf["imageCapture"])
+    save_image = SaveImage(args["output"]) if args["output"] else None
     show_image = ShowImage("Image", delay=0) if args["display"] else None
     metrics = Metrics().start()
     progress = Progress()
@@ -48,13 +49,11 @@ def capture_image(args):
                 break
 
             # Get rid of input path from filename
-            if args["input"] != filename:
+            if conf["imageCapture"]["path"] != filename:
                 # Get rid of input path from filename
-                filename = os.path.relpath(filename, start=args["input"])
+                filename = os.path.relpath(filename, start=conf["imageCapture"]["path"])
             else:
                 filename = os.path.basename(filename)
-            # Getting the name of the file without the extension
-            name = os.path.splitext(filename)[0]
 
             # Visualize data
             #
@@ -62,7 +61,7 @@ def capture_image(args):
             visualize_image_info(image, filename)
 
             if save_image:
-                save_image(image, name)
+                save_image(image, filename)
 
             if show_image:
                 show = show_image(image)
@@ -108,12 +107,15 @@ class VisualizeDataPipe:
 
 
 def capture_image_pipeline(args):
+    # Setup logger
     logger = logging.getLogger(__name__)
+    # Read configuration
+    conf = load_config(args["conf"], args["conf_overwrites"])
 
     # Setup pipeline steps
-    capture_image_pipe = CaptureImagePipe(args["input"], "." + args["image_ext"], args["contains"])
+    capture_image_pipe = CaptureImagePipe(conf["imageCapture"])
     visualize_data_pipe = VisualizeDataPipe("vis_image")
-    save_image_pipe = SaveImagePipe("vis_image", args["output"], args["image_ext"]) if args["output"] else None
+    save_image_pipe = SaveImagePipe("vis_image", args["output"]) if args["output"] else None
     show_image_pipe = ShowImagePipe("vis_image", "Video", delay=0) if args["display"] else None
     metrics_pipe = MetricsPipe()
     progress_pipe = ProgressPipe()