+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 - Possibility to overwrite YAML config settings from the command line
 - Face detection example with Dlib frontal face detector
+- YAML config for image capturing (see example/capture_image.py)
 
 # 0.1.1
 Initial release